ZMQ_HWM type changed to int

+++ src/options.hpp
@@ -35,7 +35,9 @@ namespace zmq
         int setsockopt (int option_, const void *optval_, size_t optvallen_);
         int getsockopt (int option_, void *optval_, size_t *optvallen_);
 
-        uint64_t hwm;
+        //  High-water mark for messages in pipe.
+        int hwm;
+
         uint64_t affinity;
         blob_t identity;
